nghttp2: upgrade library to g4f7223e and add casts to please buildbots
Reviewed-on: https://code.wireshark.org/review/1900

+++ wsutil/nghttp2/nghttp2_buf.c
@@ -236,8 +236,8 @@ ssize_t nghttp2_bufs_len(nghttp2_bufs *bufs)
 
 static int bufs_avail(nghttp2_bufs *bufs)
 {
-  return nghttp2_buf_avail(&bufs->cur->buf) +
-    (bufs->chunk_length - bufs->offset) * (bufs->max_chunk - bufs->chunk_used);
+  return (int)(nghttp2_buf_avail(&bufs->cur->buf) +
+    (bufs->chunk_length - bufs->offset) * (bufs->max_chunk - bufs->chunk_used));
 }
 
 static int bufs_alloc_chain(nghttp2_bufs *bufs)
@@ -424,7 +424,7 @@ ssize_t nghttp2_bufs_remove(nghttp2_bufs *bufs, uint8_t **out)
 
   *out = res;
 
-  return len;
+  return (ssize_t)len;
 }
 
 void nghttp2_bufs_reset(nghttp2_bufs *bufs)
